refactor: testutils -> internal/testutils
