+++ common/configuration/asm_test.go
@@ -17,7 +17,7 @@ import (
 	ftlv1 "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1"
 	"github.com/TBD54566975/ftl/internal/log"
 	"github.com/TBD54566975/ftl/internal/slices"
-	"github.com/TBD54566975/ftl/testutils"
+	"github.com/TBD54566975/ftl/internal/testutils"
 
 	"connectrpc.com/connect"
 	"github.com/alecthomas/assert/v2"
