+++ go-runtime/ftl/config_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/TBD54566975/ftl/go-runtime/internal"
 	"github.com/TBD54566975/ftl/internal/log"
 	"github.com/TBD54566975/ftl/internal/modulecontext"
-	. "github.com/TBD54566975/ftl/testutils/modulecontext"
+	. "github.com/TBD54566975/ftl/internal/testutils/modulecontext"
 )
 
 func TestConfig(t *testing.T) {
